add serviceworker to file serve.js
