+++ log.txt
@@ -1,0 +1,1 @@
+incidentID SC1208183237 was authenticated by BaoTM2 at 12-08-2018 02:32:47
